% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asBedMethyl.R
\name{asBedMethyl}
\alias{asBedMethyl}
\title{Dump DNA methylation values into a .bedmethyl file for e.g. MARLIN or Lamprey}
\usage{
asBedMethyl(x, BPPARAM = SerialParam(progressbar = TRUE))
}
\arguments{
\item{x}{a SummarizedExperiment-like object (DO NOT dump huge ones)}

\item{BPPARAM}{a BiocParallelParam (default is SerialParam(progressbar=T))}
}
\value{
list     a list of filenames dumped
}
\description{
Dump DNA methylation values into a .bedmethyl file for e.g. MARLIN or Lamprey
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    This function will squawk if x has more than 50 columns. 
             If x is a methBlockExperiment, fromMethBlocks is used to
             ensure that locus-level representations exist in output.
             The output is neither compressed nor tabixed at present.
             bedMethyl column spec (with notes on R implementation):
             | ---------- | --------- | ------------ |
             | *name*     | *type*    | *notes*      |
             | ---------- | --------- | ------------ |
             | chrom      | character |              |
             | ---------- | --------- | ------------ |
             | start      | integer   |              |
             | ---------- | --------- | ------------ |
             | end        | integer   |              |
             | ---------- | --------- | ------------ |
             | name       | character | m or h       |
             | ---------- | --------- | ------------ |
             | score      | integer   | 0-1000, pval |
             | ---------- | --------- | ------------ |
             | strand     | character | +/-/.        |
             | ---------- | --------- | ------------ |
             | thickStart | integer   |              |
             | ---------- | --------- | ------------ |
             | thickEnd   | integer   |              |
             | ---------- | --------- | ------------ |
             | color      | character | '255,0,0'    |
             | ---------- | --------- | ------------ |
             | nValidCov  | character | set to 100   |
             | ---------- | --------- | ------------ |
             | percMod    | numeric   | 0-100\%       |
             | ---------- | --------- | ------------ |
             | nMod       | integer   | methylated   |
             | ---------- | --------- | ------------ |
             | nCanon     | integer   | unmethlyated |
             | ---------- | --------- | ------------ |
             | nOther     | integer   | other calls  |
             | ---------- | --------- | ------------ |
             | nDelete    | integer   | reads w/dels |
             | ---------- | --------- | ------------ |
             | nFail      | integer   | failed calls |
             | ---------- | --------- | ------------ |
             | nDiff      | integer   | variant base |
             | ---------- | --------- | ------------ |
             | nNoCall    | integer   | ref nocalls  |
             | ---------- | --------- | ------------ |
             For more, consult the [bedMethyl description](https://genome.ucsc.edu/goldenpath/help/bedMethyl.html).
}\if{html}{\out{</div>}}
}
