% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asMethBlocks.R
\name{asMethBlocks}
\alias{asMethBlocks}
\title{Collapse a GenomicRatioSet or similar over methylation blocks (per Kaplan)}
\usage{
asMethBlocks(x, g = c("hg19", "hg38", "mm10", "custom"), custom = NULL)
}
\arguments{
\item{x}{a SummarizedExperiment-derived object with CpG loci as rows}

\item{g}{genome to use for methylation block coordinates (hg19)}

\item{custom}{if genome == "custom", a GRanges (cannot be NULL) of blocks}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     an object with same colData but with new rowRanges & assays
}\if{html}{\out{</div>}}
}
\description{
By default, data(methBlocks) is used for summarization, singleton probes are
included, methylation block rownames are returned against hg19, and neither
singletons (!proper) nor unstable mappings (!stable) are filtered out. This
version of asMethBlocks supports "custom" as a target; be careful with that.
}
\details{
If you have specified a genome() for x, which you very much should, this
function will raise an error if the genomes or seqlevels do not match. A copy
of the subset of methBlocks used for mapping will be returned in metadata(x).
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    switchMethBlocksGenome
}\if{html}{\out{</div>}}
}
