% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignBlocks.R
\name{assignBlocks}
\alias{assignBlocks}
\title{assign loci to regions (blocks) using standard GenomicRanges tools}
\usage{
assignBlocks(loci, blocks, fill = FALSE)
}
\arguments{
\item{loci}{a GRanges of (usually smaller) regions}

\item{blocks}{a GRanges of (usually larger) regions}

\item{fill}{assign unmatched loci to their own blocks? (FALSE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     the loci but with mcols(loci)$block filled with block names
}\if{html}{\out{</div>}}
}
\description{
This is a somewhat trivial function that I got tired of writing over & over
}
\details{
If you have specified a genome() for loci and blocks, which you ought to,
this function will raise an error if the genomes or seqlevels do not match.
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    asMethBlocks
}\if{html}{\out{</div>}}
}
