% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBlocksToMBE.R
\name{betaBlocksToMBE}
\alias{betaBlocksToMBE}
\title{read beta blocks (wgbstools beta_to_table output) into a MethBlockExperiment}
\usage{
betaBlocksToMBE(tbx, n = -1, g = "hg19")
}
\arguments{
\item{tbx}{either a TabixFile or something that can be turned into one}

\item{n}{how many lines to read? (default is -1, i.e., all of them)}

\item{g}{optional genome indicator (default is NULL)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ a MethBlockExperiment
}\if{html}{\out{</div>}}
}
\description{
read beta blocks (wgbstools beta_to_table output) into a MethBlockExperiment
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{realistically, this could be done with scanTabix, I just don't. 
}\if{html}{\out{</div>}}
}
