% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockDownsample.R
\name{blockDownsample}
\alias{blockDownsample}
\title{block-randomized sampling to ensure somewhat representative comparisons}
\usage{
blockDownsample(x, col, maxN = Inf, SNPcall = FALSE)
}
\arguments{
\item{x}{a MethylationExperiment or comparable rectangular object}

\item{col}{a factor for each column, or its name in colData}

\item{maxN}{maximum size per-subset? (Inf; use the smallest subset size)}

\item{SNPcall}{recall SNPs if present? (FALSE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a subset of x, downsampled to match the smallest N(col)
}\if{html}{\out{</div>}}
}
\description{
block-randomized sampling to ensure somewhat representative comparisons
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  use set.seed beforehand if you want reproducibility.
}\if{html}{\out{</div>}}
}
