% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byExtremality.R
\name{byExtremality}
\alias{byExtremality}
\title{byExtremality}
\usage{
byExtremality(x, k = 500)
}
\arguments{
\item{x}{a matrix of beta values (proportions), by block or by locus}

\item{k}{how many rows to return (500)}
}
\value{
the most extremal \emph{k} rows of \emph{x}
}
\description{
For array-based DNA methylation, particularly rates across regions,
we can do better (a lot better) than MAD. An ideal class-splitting Beta-
distributed feature will have Bernoulli variance (== all 0s and 1s):
}
\details{
max(SD(X_j)) if X_j ~ Beta(a, b) <= max(SD(X_j)) if X_j ~ Bernoulli(a/(a+b))

for X having a known mean and SD, solvable for a + b by MoM. Now define

extremality(X_j) = sd(X_j) / bernoulliSD(mean(X_j))

This function selects the k most extremal rows of x and returns their values.
}
