% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bySD.R
\name{bySD}
\alias{bySD}
\title{bySD}
\usage{
bySD(x, k = 100, minmean = 0.025, maxmean = 0.975)
}
\arguments{
\item{x}{a numeric matrix, possibly with NAs}

\item{k}{maximum number of rows to return (100)}

\item{minmean}{minimum mean value (0.025)}

\item{maxmean}{maximum mean value (0.975)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   the most variable rows of x (with suitable means)
}\if{html}{\out{</div>}}
}
\description{
This function selects the k most variable rows of x and returns their values.
}
