% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorWeight.R
\name{factorWeight}
\alias{factorWeight}
\title{extract the positional weights for an NMF factor}
\usage{
factorWeight(x, y = 1, g = NULL)
}
\arguments{
\item{x}{a MethylationBlockExperiment, usually}

\item{y}{which factor (1)}

\item{g}{which genome (whichever one x claims to be using)}
}
\value{
a GRanges where the $score column is that factor's weights
}
\description{
extract the positional weights for an NMF factor
}
\details{
like duh, if there's no mcols(x)[\link{"NMF"}], this won't work.
You will want to use rtracklayer or igvR to view the result;
seqinfo(gr) <- SeqinfoForUCSCGenome(g)\link{seqlevels(gr)} may help
}
