% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByMean.R
\name{filterByMean}
\alias{filterByMean}
\title{filterByMean}
\usage{
filterByMean(x, min = 0.025, max = 0.975, INDEX = c(1, 2))
}
\arguments{
\item{x}{a matrix to filter}

\item{min}{minimum allowable mean (0.025)}

\item{max}{maximum allowable mean (0.975)}

\item{INDEX}{which dimension to filter along (1)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ only the permissible (rows/columns) of X
}\if{html}{\out{</div>}}
}
\description{
does what it says
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{this could probably be adapted to operate on INDEX=1:2
}\if{html}{\out{</div>}}
}
