% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNmfAndUmap.R
\name{fitNmfAndUmap}
\alias{fitNmfAndUmap}
\title{performs NMF on the beta values, stuffs the factors into colData, runs UMAP}
\usage{
fitNmfAndUmap(SCE, k = 20, ...)
}
\arguments{
\item{SCE}{a SingleCellExperiment (or SummarizedExperiment to coerce to one)}

\item{k}{number of factors to find (20)}

\item{...}{other arguments to pass to RcppML::nmf_fit}
}
\value{
SCE, with new reducedDim()s, colData, nmf_fit & umap_fit metadata
}
\description{
performs NMF on the beta values, stuffs the factors into colData, runs UMAP
}
