% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromMethBlocks.R
\name{fromMethBlocks}
\alias{fromMethBlocks}
\title{extract methylation rates at probes corresponding to methylation blocks and
expand the matrix to match the original input dimensions, for (e.g.) data
transfer or comparison across platforms and technologies (arrays, ONT, etc.)}
\usage{
fromMethBlocks(x, loci)
}
\arguments{
\item{x}{a GenomicRatioSet or something like it with chr12:34-56 rows}

\item{loci}{probe coordinates (or a GRanges) to unpack blocks onto}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     an object reinflated to locus level using rates
}\if{html}{\out{</div>}}
}
\description{
see documentation for asMethBlocks and toMethBlocks for more details
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    switchMethBlocksGenome
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{    toMethBlocks
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{    asMethBlocks
}\if{html}{\out{</div>}}
}
