% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.GimmeCpG.R
\name{impute.GimmeCpG}
\alias{impute.GimmeCpG}
\title{impute missing sites via GimmeCpG (binned within 35bp, weighted out to 1kbp)}
\usage{
impute.GimmeCpG(x, g = NULL, BPPARAM = SerialParam(progressbar = TRUE))
}
\arguments{
\item{x}{a MethylationExperiment (or something with getBeta and coords)}

\item{g}{a genome (hg19 is the default, or use coords from granges(x))}

\item{BPPARAM}{a BiocParallelParam object (default SerialParam(progr=TRUE))}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   x, but with a new assay Imputed that has been imputed.
}\if{html}{\out{</div>}}
}
\description{
impute missing sites via GimmeCpG (binned within 35bp, weighted out to 1kbp)
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  this function has not yet been tested enough for prime time.
}\if{html}{\out{</div>}}
}
