% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeByClass.R
\name{imputeByClass}
\alias{imputeByClass}
\title{impute missing sites via k-NN with user-defined classes for comparables}
\usage{
imputeByClass(x, col, maxK = 10, noise = 0, ..., BPPARAM = NULL)
}
\arguments{
\item{x}{a MethylationExperiment (or something with getBeta)}

\item{col}{a factor with the class of each column, or its name in colData}

\item{maxK}{maximum K (positive integer) nearest neighbors to use (10)}

\item{noise}{add noise to observations? (0; noise > 0 can help with uWGBS)}

\item{...}{other arguments to pass along to impute::impute.knn}

\item{BPPARAM}{a BiocParallel::MulticoreParam() or similar for processing}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   x, with an assay named RawBetas holding the originals
}\if{html}{\out{</div>}}
}
\description{
impute missing sites via k-NN with user-defined classes for comparables
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  in principle, this could be combined with imputeGimmeCpG;
           in practice, can also iterate with coarsening 'col' factors.
           for sequencing-based assays, setting 'noise' > 0 might help. 
}\if{html}{\out{</div>}}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  impute::impute.knn
}\if{html}{\out{</div>}}
}
