% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeGimmeCpG.R
\name{imputeGimmeCpG}
\alias{imputeGimmeCpG}
\title{impute missing sites via GimmeCpG (binned within 35bp, weighted out to 1kbp)}
\usage{
imputeGimmeCpG(x, g = NULL, BPPARAM = SerialParam(progressbar = TRUE))
}
\arguments{
\item{x}{a MethylationExperiment (or something with getBeta and coords)}

\item{g}{a genome (hg19 is the default, or use coords from granges(x))}

\item{BPPARAM}{a BiocParallelParam object (default SerialParam(progr=TRUE))}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   x, with an additional assay RawBeta if none currently exists
}\if{html}{\out{</div>}}
}
\description{
impute missing sites via GimmeCpG (binned within 35bp, weighted out to 1kbp)
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  this function has not yet been tested enough for prime time.
           once it has, it may be useful to combine with imputeByClass,
           and one may want to investigate settings for 'noise' therein.
}\if{html}{\out{</div>}}
}
