% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethBlockExperiment.R
\name{MethBlockExperiment}
\alias{MethBlockExperiment}
\alias{genome<-,MethBlockExperiment-method}
\title{The MethBlockExperiment class (inherits from ME, which inherits from SCE).}
\usage{
MethBlockExperiment(x, ...)

\S4method{genome}{MethBlockExperiment}(x) <- value
}
\arguments{
\item{x}{a MethylationExperiment}

\item{...}{additional arguments to asMethBlocks}
}
\value{
a MethBlockExperiment
}
\description{
The difference between a MethBlockExperiment and a MethylationExperiment is
that the former (this class) represents mappings(blocks) of correlated
DNA methylation regions for at least one genome assembly (usually 2 or 3).
Since methylation rates tend to be more stable across regions than at
individual loci, this can be useful for low-input or cross-platform
comparison. On the other hand, if one wants to model 'epi-clones' or
similar stochastic process based representations, a MethylationExperiment
or perhaps bsseq or epiallele representation may be preferable. Use the
right tool for the experimental question at hand. We find that methylation
blocks are often the right tool for modeling cell states and fates.
}
\details{
Note that genome(MBE) <- 'hg38' or similar will attempt remapping of rows.
If this fails, the original (un-remapped) object will be returned instead.

\if{html}{\out{<div class="sourceCode">}}\preformatted{  Can go straight from IDATs to blocks by chaining this. 
}\if{html}{\out{</div>}}
}
