% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationExperiment.R
\name{MethylationExperiment}
\alias{MethylationExperiment}
\title{MethylationExperiment() constructor function}
\usage{
MethylationExperiment(...)
}
\arguments{
\item{...}{arguments passed to the SingleCellExperiment() constructor}
}
\value{
a MethylationExperiment
}
\description{
MethylationExperiment() constructor function
}
\details{
If openSesameToME is used, a matrix of total intensities can
be added to assays(object, "CN") with per-probe intensities.
SingleCellExperiment::splitAltExps separates CpG, CpH & SNP probes.
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{openSesameToME
}\if{html}{\out{</div>}}
}
