% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{missingness,ANY,missing,missing-method}
\alias{missingness,ANY,missing,missing-method}
\title{missingness()}
\usage{
\S4method{missingness}{ANY,missing,missing}(X, MARGIN, i, ...)
}
\arguments{
\item{X}{a rectangular data structure}

\item{MARGIN}{which margin to tally on (1=rows, 2=columns) (2)}

\item{i}{which assay to tally, if X is a SummarizedExperiment}

\item{...}{other parameters, currently ignored}
}
\description{
What fraction of a rectangular data structure is NA, by row or by column?
}
\examples{
 

vals <- runif(n=100)
is.na(vals) <- as.logical(rbinom(length(vals), size=1, prob=0.1))
mat <- matrix(vals, ncol=20, nrow=5)
colnames(mat) <- paste0("column", seq_len(ncol(mat)))
rownames(mat) <- paste0("row", seq_len(nrow(mat)))
missingness(mat)
missingness(mat, 1)

}
\seealso{
applySCE

rowSums

colSums

is.na
}
