% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSNPcalls.R
\name{plotSNPcalls}
\alias{plotSNPcalls}
\title{call and plot SNPs from a matrix of SNP betas}
\usage{
plotSNPcalls(
  x,
  rotate = FALSE,
  qc = FALSE,
  BPPARAM = NULL,
  pal = c("jet", "bw"),
  ...
)
}
\arguments{
\item{x}{an object with SNPs in its metadata(), or a SNP matrix}

\item{rotate}{rotate the subjects onto the side? (FALSE)}

\item{qc}{QC on the SNPs? (implies rotate) (FALSE)}

\item{BPPARAM}{a BiocParallelParam() object, or (default) SerialParam()}

\item{pal}{palette to use ("jet" or "bw") ("jet")}

\item{...}{other arguments passed on to Heatmap}
}
\description{
plot SNPs living in metadata(x)$SNPs, provided they match sampleNames(x)
}
\details{
Plotting is done by .plotSNPcallmat() and .plotSNPheat()
If SNP calls are found in metadata(x)$SNPcalls, they will be used,
otherwise SNPs will be called on the fly (which can be sloooow).
}
