% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectNewData.R
\name{projectNewData}
\alias{projectNewData}
\title{project new data into an existing NMF model using RcppML::project}
\usage{
projectNewData(x, y, how = c("both", "nmf", "umap"), ...)
}
\arguments{
\item{x}{a MethBlockExperiment, SingleCellExperiment, or nmf object (query)}

\item{y}{a MethBlockExperiment, SingleCellExperiment, or nmf object (model)}

\item{how}{how to project new data into the existing space? (both, nmf, umap)}

\item{...}{other arguments to pass to RcppML::project and/or uwot::umap}
}
\value{
a ReducedExperiment or FactorisedExperiment (see Details)
}
\description{
project new data into an existing NMF model using RcppML::project
}
\details{
the type of returned object depends on "how": if "both" (the
default value), a ReducedExperiment with UMAP reduction will be
returned with a FactorisedExperiment in metadata()$nmf_projection.
This provides both a UMAP projection and the NMF W and H matrices.
If how == "nmf", only the FactorisedExperiment will be returned.
If how == "umap", only the ReducedExperiment will be returned.
If we start supporting more methods, we'll figure something out.
}
