% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBetaBlocks.R
\name{readBetaBlocks}
\alias{readBetaBlocks}
\title{read a tabix'ed beta block file from wgbstools, where columns 1-5 are coords}
\usage{
readBetaBlocks(tbx, n = -1, g = NULL)
}
\arguments{
\item{tbx}{either a TabixFile or something that can be turned into one}

\item{n}{how many lines to read? (default is -1, i.e., all of them)}

\item{g}{optional genome indicator (default is NULL)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ a matrix, possibly with a GRanges in attr(, "GRanges")
}\if{html}{\out{</div>}}
}
\description{
read a tabix'ed beta block file from wgbstools, where columns 1-5 are coords
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{realistically, this could be done with scanTabix, I just don't. 
}\if{html}{\out{</div>}}
}
