% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBetaFile.R
\name{readBetaFile}
\alias{readBetaFile}
\title{read wgbstools .beta file into a matrix (CpG bed into attr(, "GRanges"))}
\usage{
readBetaFile(x, CpG = NULL, g = NULL)
}
\arguments{
\item{x}{path to the file}

\item{CpG}{optional but HIGHLY recommended CpG dictionary for x (NULL)}

\item{g}{optional but HIGHLY recommended genome indicator (NULL)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ a matrix of methylated (column 1) and total (column 2) reads
}\if{html}{\out{</div>}}
}
\description{
read wgbstools .beta file into a matrix (CpG bed into attr(, "GRanges"))
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{the output is nearly useless without the CpG dictionary. If the
         size of the file is 56434896, we know g == "hg19", but the CpG
         dictionary for hg19 is 125MB, so we don't have it in data(). 
}\if{html}{\out{</div>}}
}
