% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducedData.R
\name{reducedData}
\alias{reducedData}
\title{grab a data.frame of a reducedDim representation, name columns, bind colData}
\usage{
reducedData(SCE, name, withcd = TRUE)
}
\arguments{
\item{SCE}{a SingleCellExperiment or something that acts like one}

\item{name}{the name of a reducedDim in SCE (will be used for names)}

\item{withcd}{bind colData to the result? (TRUE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a data.frame for feeding to ggplot2 or similar
}\if{html}{\out{</div>}}
}
\description{
grab a data.frame of a reducedDim representation, name columns, bind colData
}
