% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitUmap.R
\name{refitUmap}
\alias{refitUmap}
\title{refit a UMAP on a subset of NMF columns and stuff it in new reducedDim}
\usage{
refitUmap(SCE, k_sub, name)
}
\arguments{
\item{k_sub}{which NMF factors to use for recomputing the UMAP}

\item{name}{add the weights to rowData(x)? (TRUE)}

\item{x}{a MethBlockExperiment or similar thing-with-Beta-values}

\item{...}{other arguments to pass to umap}
}
\value{
SCE, with new reducedDim()s, and \code{name} metadata (model)
}
\description{
refit a UMAP on a subset of NMF columns and stuff it in new reducedDim
}
\details{
the RSpectra import is to avoid a weird umap bug.
}
