% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeeze.R
\name{squeeze}
\alias{squeeze}
\alias{unsqueeze}
\title{squeeze a proportion by a finite amount}
\usage{
squeeze(p, sqz = 1e-06, min = 0, max = 1)
}
\arguments{
\item{p}{a vector of proportions (for squeeze)}

\item{sqz}{a squeeze factor (1e-6, by default)}

\item{min}{minimum unsqueezed value possible (0)}

\item{max}{maximum unsqueezed value possible (1)}

\item{x}{a vector of real numbers (for unsqueeze)}
}
\value{
(0 + sqz) <= p <= (1 - sqz), or for unsqueeze, min <= p <= max.
}
\description{
squeeze a proportion by a finite amount
}
\details{
squeeze and unsqueeze avoid -Inf, +Inf, and NaNs in transforms.
}
\examples{

expit <- function(x) exp(x) / (1 + exp(x))
logit <- function(p) log(p / (1 - p))

set.seed(1234)
x <- rnorm(n=1e6)
p <- round(expit(x), 2)
summary(logit(p) - x)

set.seed(12345)
x <- rnorm(n=1e6)
p <- round(expit(x), 2) 
summary(logit(p) - x)          # oopsie!
summary(logit(squeeze(p)) - x) # better!

}
