% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchMethBlocksGenome.R
\name{switchMethBlocksGenome}
\alias{switchMethBlocksGenome}
\title{Given an object with methylation block coordinates, switch the genome}
\usage{
switchMethBlocksGenome(x, to = c("hg38", "hg19"))
}
\arguments{
\item{x}{a GenomicRatioSet or something like it with methBlock coords}

\item{to}{the new genome to use (hg19 or hg38)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     an object very much like x, but mapped onto `to`
}\if{html}{\out{</div>}}
}
\description{
Given an object with methylation block coordinates, switch the genome
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    asMethBlocks
}\if{html}{\out{</div>}}
}
