% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toMethBlocks.R
\name{toMethBlocks}
\alias{toMethBlocks}
\title{extract the probes corresponding to methylation blocks and collapse them}
\usage{
toMethBlocks(x, y, g = c("hg19", "hg38"), methBlocks = NULL)
}
\arguments{
\item{x}{a GenomicRatioSet or something like it with cgXXXXXXXXX rows}

\item{y}{block coordinates (or a GRanges) to summarize x over}

\item{g}{genome to use for block coordinates if none found (hg19)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     an object or matrix summarized down to the methBlocks in y
}\if{html}{\out{</div>}}
}
\description{
see documentation for asMethBlocks and fromMethBlocks for more details
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    switchMethBlocksGenome
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{    fromMethBlocks
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{    asMethBlocks
}\if{html}{\out{</div>}}
}
