% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unVST.R
\name{unVST}
\alias{unVST}
\title{Inverse variance stabilizing transform (sin(x)**2) for a proportion}
\usage{
unVST(x)
}
\arguments{
\item{x}{a vector of transformed proportions (as from VST(p))}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ values between 0 and 1
}\if{html}{\out{</div>}}
}
\description{
Inverse variance stabilizing transform (sin(x)**2) for a proportion
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Using asin(sqrt(p)) for VST avoids negatives and infinities.
}\if{html}{\out{</div>}}
}
\examples{

dat <- rbeta(n=100, 1, 7)
vsted <- VST(dat)
all(abs(unVST(vsted) - dat) < .Machine$double.eps)

}
\seealso{
VST

extremality
}
