% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VST.R
\name{VST}
\alias{VST}
\title{Variance stabilizing transform (asin(sqrt(p))) for a proportion}
\usage{
VST(p, min = 0, max = 1)
}
\arguments{
\item{p}{a vector of proportions (VST)}

\item{min}{minimum valid input (0)}

\item{max}{maximum valid input (1)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ values between 0 and 1.570796327
}\if{html}{\out{</div>}}
}
\description{
Variance stabilizing transform (asin(sqrt(p))) for a proportion
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Using asin(sqrt(p)) for VST avoids negatives and infinities.
}\if{html}{\out{</div>}}
}
\examples{

dat <- rbeta(n=100, 1, 7)
vsted <- VST(dat)
all(abs(unVST(vsted) - dat) < .Machine$double.eps)

}
\seealso{
unVST

extremality
}
