% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightsBy.R
\name{weightsBy}
\alias{weightsBy}
\title{plot the relative representation of some grouping by reducedData factor}
\usage{
weightsBy(SCE, by, name = "NMF", add = NULL)
}
\arguments{
\item{SCE}{something that quacks like a SingleCellExperiment}

\item{by}{name of a variable (usually a factor) in colData}

\item{name}{name of a reducedDim in SCE ("NMF")}

\item{add}{optional confounder to include for e.g. faceting (NULL)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ a ggplot object
}\if{html}{\out{</div>}}
}
\description{
plot the relative representation of some grouping by reducedData factor
}
