% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byExtremality.R
\name{byExtremality}
\alias{byExtremality}
\title{byExtremality}
\usage{
byExtremality(x, k = 100, minmean = 0.025, maxmean = 0.975)
}
\arguments{
\item{x}{a matrix of beta values (proportions), by block or by locus}

\item{k}{how many rows to return (100)}

\item{minmean}{minimum row mean (0.05)}

\item{maxmean}{maximum row mean (0.95)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   the most extremal _k_ rows of _x_ (with suitable means)
}\if{html}{\out{</div>}}
}
\description{
For array-based DNA methylation, particularly rates across regions,
we can do better (a lot better) than MAD. An ideal class-splitting Beta-
distributed feature will have Bernoulli variance (== all 0s and 1s):
}
\details{
max(SD(X_j)) if X_j ~ Beta(a, b) <= max(SD(X_j)) if X_j ~ Bernoulli(a/(a+b))

for X having a known mean and SD, solvable for a + b by MoM. Now define

extremality(X_j) = sd(X_j) / bernoulliSD(mean(X_j))

This function selects the k most extremal rows of x and returns their values.

\if{html}{\out{<div class="sourceCode">}}\preformatted{  it turns out that you usually want 0.975 >= mean >= 0.025
}\if{html}{\out{</div>}}
}
