% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bySD.R
\name{bySD}
\alias{bySD}
\title{bySD}
\usage{
bySD(x, k = 500, minmean = -Inf, maxmean = Inf)
}
\arguments{
\item{x}{a numeric matrix, possibly with NAs}

\item{k}{how many rows to return (500)}

\item{minmean}{minimum mean value (-Inf)}

\item{maxmean}{maximum mean value (+Inf)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   the most variable _k_ rows of _x_ (with suitable means)
}\if{html}{\out{</div>}}
}
\description{
This function selects the k most variable rows of x and returns their values.
}
