% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorWeight.R
\name{factorWeight}
\alias{factorWeight}
\title{extract the positional weights for an NMF factor}
\usage{
factorWeight(x, y = 1, g = NULL)
}
\arguments{
\item{x}{a MethylationBlockExperiment, usually}

\item{y}{which factor (1)}

\item{g}{which genome (whichever one x claims to be using)}
}
\value{
a GRanges where the $score column is that factor's weights
}
\description{
extract the positional weights for an NMF factor
}
\details{
If mcols(x) does not contain a matrix called NMF, this will fail.
Also, you'll want to use rtracklayer or igvR to view the result.
(see example for a demonstration of how this approach can work)
}
\examples{
\dontrun{
  nmf5 <- factorWeight(MBE, 5)
  nmf7 <- factorWeight(MBE, 7) 
  library(rtracklayer) 
  for (i in c("nmf5","nmf7")) {
    j <- get(i)
    seqinfo(j) <- SeqinfoForUCSCGenome(unique(genome(j)))[seqlevels(j)]
    export(j, paste(i, unique(genome(j)), "bw", sep="."))
  }
}

}
