% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNmfAndUmap.R
\name{fitNmfAndUmap}
\alias{fitNmfAndUmap}
\title{performs NMF on the beta values, stuffs the factors into colData, runs UMAP}
\usage{
fitNmfAndUmap(SCE, k = 30, ...)
}
\arguments{
\item{k}{number of factors to find (30) (FIXME: use ARD instead)}

\item{...}{other arguments to pass to RcppML::nmf}

\item{x}{a MethBlockExperiment or similar thing-with-Beta-values}
}
\value{
SCE, with new reducedDim()s, colData, nmf_fit & umap_fit metadata
}
\description{
performs NMF on the beta values, stuffs the factors into colData, runs UMAP
}
\details{
the RSpectra import is to avoid a weird umap bug.
}
