% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.classKNN.R
\name{impute.classKNN}
\alias{impute.classKNN}
\title{impute missing sites via k-NN with user-defined classes for comparables}
\usage{
impute.classKNN(x, col, ...)
}
\arguments{
\item{x}{a MethylationExperiment (or something with getBeta)}

\item{col}{a factor with the class of each column, or its name in colData}

\item{...}{other arguments to pass along to impute::impute.knn}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   x, but with a new assay named ImputedBeta 
}\if{html}{\out{</div>}}
}
\description{
impute missing sites via k-NN with user-defined classes for comparables
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  in principle, this can be combined with impute.GimmeCpG
}\if{html}{\out{</div>}}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  impute::impute.knn
}\if{html}{\out{</div>}}
}
