% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.classKNN.R
\name{impute.classKNN}
\alias{impute.classKNN}
\title{impute missing sites via k-NN with user-defined classes for comparables}
\usage{
impute.classKNN(X, INDEX, ...)
}
\arguments{
\item{X}{a MethylationExperiment (or something with getBeta)}

\item{INDEX}{a factor representing the class of each observation}

\item{...}{other arguments to pass along to impute::impute.knn}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   x, but with a new assay Imputed that has been imputed.
}\if{html}{\out{</div>}}
}
\description{
impute missing sites via k-NN with user-defined classes for comparables
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  in principle, this can be combined with impute.GimmeCpG
}\if{html}{\out{</div>}}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  impute::impute.knn
}\if{html}{\out{</div>}}
}
