% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.classKNN.R
\name{impute.classKNN}
\alias{impute.classKNN}
\title{impute missing sites via k-NN with user-defined classes for comparables}
\usage{
impute.classKNN(x, col, maxK = 10, BPPARAM = NULL, ...)
}
\arguments{
\item{x}{a MethylationExperiment (or something with getBeta)}

\item{col}{a factor with the class of each column, or its name in colData}

\item{maxK}{maximum K (positive integer) nearest neighbors to use (10)}

\item{BPPARAM}{a BiocParallel::MulticoreParam() or similar for processing}

\item{...}{other arguments to pass along to impute::impute.knn}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   x, but with an assay named RawBetas holding the originals
}\if{html}{\out{</div>}}
}
\description{
impute missing sites via k-NN with user-defined classes for comparables
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  in principle, this could be combined with impute.GimmeCpG
           in practice, can iterate this with coarsening 'col' factors
}\if{html}{\out{</div>}}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  impute::impute.knn
}\if{html}{\out{</div>}}
}
