% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEME.R
\name{iSEEME}
\alias{iSEEME}
\title{convenience wrapper for iSEE on a MethylationExperiment}
\usage{
iSEEME(x, colorColumn = "cell.type")
}
\arguments{
\item{x}{a MethylationExperiment with NMF and UMAP reducedDims}

\item{colorColumn}{name of the colData column to color plots ("cell.type")}
}
\description{
Note that this is a quick and dirty affair so don't expect much.
}
\details{
presumes NMF and UMAP. you will almost certainly want to tweak this.
}
