% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markMergeable.R
\name{markMergeable}
\alias{markMergeable}
\title{find mergeable methylation blocks by correlation (raw or by NMF weights)}
\usage{
markMergeable(x, how = c("r", "nmf"), thr = 0.9)
}
\arguments{
\item{x}{a MethBlockExperiment, perhaps with "nmf_fit" in metadata}

\item{how}{correlation ('r') or NMF weight correlation ('nmf')}

\item{thr}{threshold for merging neighboring blocks (0.9; see Details)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ a list, GRanges, or MBE of merged blocks based on (how, thr)
}\if{html}{\out{</div>}}
}
\description{
find mergeable methylation blocks by correlation (raw or by NMF weights)
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{blocks more than max(width(blocks)) bp apart are not checked
}\if{html}{\out{</div>}}
}
