% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotME.R
\name{plotME}
\alias{plotME}
\title{plotME}
\usage{
plotME(
  x,
  k = 100,
  minmean = 0.2,
  maxmean = 0.8,
  maxNA = 0.2,
  splitBy = NULL,
  ...,
  BPPARAM = NULL
)
}
\arguments{
\item{x}{a MethylationExperiment or MethBlockExperiment}

\item{k}{how many features to use (100)}

\item{minmean}{minimum mean for byExtremality (0.2)}

\item{maxmean}{maximum mean for byExtremality (0.8)}

\item{maxNA}{maximum fraction NA for a feature (0.2)}

\item{splitBy}{a column name on which to split rows (NULL)}

\item{...}{parameters to pass to Heatmap}

\item{BPPARAM}{BiocParallelParam() to pass to plotSNPcalls (SerialParam())}
}
\description{
plot a MethylationExperiment or MethBlockExperiment features + SNPs
}
\details{
byExtremality() is called to determine features. A joint plot
of highly variable features on the left and SNPs on the right
is returned. This function is purely for a convenient first pass.
Recent experiences with low-coverage WGBS have convinced us that
bounding minimum and maximum mean is important for byExtremality.
If SNPs cannot be found, only the main methylation plot is returned.
For reasons not yet clear, it's usually best to leave BPPARAM alone.
For reasons quite clear, caching SNP calls is a very good idea, via
metadata(x)$SNPcalls <- SNPcalls(x) # will speed up things A LOT
}
