% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotME.R
\name{plotME}
\alias{plotME}
\title{plotME}
\usage{
plotME(x, k = 100, minmean = 0.2, maxmean = 0.8, ...)
}
\arguments{
\item{x}{a MethylationExperiment or MethBlockExperiment}

\item{k}{how many features to use (100)}

\item{minmean}{minimum mean for byExtremality (0.2)}

\item{maxmean}{maximum mean for byExtremality (0.8)}

\item{...}{parameters to pass to Heatmap}
}
\description{
plot a MethylationExperiment or MethBlockExperiment features + SNPs
}
\details{
byExtremality() is called to determine features. A joint plot
of highly variable features on the left and SNPs on the right
is returned. This function is purely for a convenient first pass.
Recent experiences with low-coverage WGBS have convinced us that
bounding minimum and maximum mean is important for byExtremality.
}
