% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotME.R
\name{plotME}
\alias{plotME}
\title{plotME}
\usage{
plotME(x, k = 500, ...)
}
\arguments{
\item{x}{a MethylationExperiment or MethBlockExperiment}

\item{k}{how many features to use (500)}
}
\description{
plot a MethylationExperiment or MethBlockExperiment features + SNPs
}
\details{
byExtremality() is called to determine features. A joint plot
of highly variable features on the left and SNPs on the right
is returned. This function is purely for a convenient first pass.
}
